code/pyproject.toml
[tool.poetry]
name = "synthetic-patients"
version = "0"
description = "na"
authors = ["Alex Goodell <alexgoodell@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.16"
langchain-openai = "^0.1.4"
load-dotenv = "^0.1.0"
jupyter = "^1.0.0"
tabulate = "^0.9.0"
rich = "^13.7.1"
pandas = "^2.2.2"
langchain-community = "^0.0.34"
ipython = "^8.24.0"
flask = "^3.0.3"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
librosa = "0.7.0"
soundfile = "^0.12.1"
faster-whisper = "^1.0.1"
opencv-python = "4.8.0.76"
torch = "2.0.1"
tqdm = "4.66.1"
moviepy = "1.0.3"
elevenlabs = "^1.2.1"
sounddevice = "^0.4.6"
numpy = "^1.26.4"
keyboard = "^0.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
